Fixed bug #767

+++ cpp/src/IceBox/ServiceManagerI.cpp
@@ -7,6 +7,7 @@
 //
 // **********************************************************************
 
+#include <IceUtil/DisableWarnings.h>
 #include <Ice/Ice.h>
 #include <Ice/DynamicLibrary.h>
 #include <Ice/SliceChecksums.h>
@@ -344,7 +345,7 @@ IceBox::ServiceManagerI::start(const string& service, const string& entryPoint, 
 	    // communicator with argc/argv. This is necessary for Ice
 	    // plugin properties (e.g.: IceSSL).
 	    //
-	    int argc = serviceArgs.size();
+	    int argc = static_cast<int>(serviceArgs.size());
 	    char** argv = new char*[argc + 1];
 	    int i = 0;
 	    for(Ice::StringSeq::const_iterator p = serviceArgs.begin(); p != serviceArgs.end(); ++p, ++i)
